+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/disabled/AutoRedA.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Disabled;
+package org.firstinspires.ftc.teamcode.disabled;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
@@ -11,24 +11,24 @@ import org.firstinspires.ftc.teamcode.toolkit.ULLinearOpMode;
 import java.util.ArrayList;
 
 @Disabled
-@Autonomous(name = "AutoBlueB", group= "Opmodes" )
+@Autonomous(name = "AutoRedA", group = "OpModes")
 
-public class AutoBlueB extends ULLinearOpMode {
+public class AutoRedA extends ULLinearOpMode {
 
     Robot robot;
     Odometry odom;
 
     @Override
-    public void runOpMode(){
-        robot = new Robot();
-        odom = new Odometry(robot);
+    public void runOpMode() {
+        robot = new Robot ();
+        odom = new Odometry (robot);
 
         waitForStart();
 
         ArrayList<PathPoint> allPoints = new ArrayList<>();
 
-        allPoints.add(new PathPoint(0, 86, 0.7, 2, 5));
-       //allPoints.add(new PathPoint(-20, 86, 0.3, 2, 5));
+        allPoints.add(new PathPoint(0, 72, 0.7, 2, 5));
+        //allPoints.add(new PathPoint(38, 72, 0.5, 2, 5));
 
         odom.followPath(allPoints);
