+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/disabled/AutoRedC.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Disabled;
+package org.firstinspires.ftc.teamcode.disabled;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
@@ -11,9 +11,9 @@ import org.firstinspires.ftc.teamcode.toolkit.ULLinearOpMode;
 import java.util.ArrayList;
 
 @Disabled
-@Autonomous(name = "AutoBlueC", group = "Opmodes")
+@Autonomous(name = "AutoRedC", group = "OpModes")
 
-public class AutoBlueC extends ULLinearOpMode {
+public class AutoRedC extends ULLinearOpMode {
 
     Robot robot;
     Odometry odom;
@@ -28,7 +28,7 @@ public class AutoBlueC extends ULLinearOpMode {
         ArrayList<PathPoint> allPoints = new ArrayList<>();
 
         allPoints.add(new PathPoint(0, 124, 0.7, 2, 5));
-        //allPoints.add(new PathPoint(-42, 124, 0.5, 2, 5));
+        //allPoints.add(new PathPoint(42, 124, 0.5, 2, 5));
 
         odom.followPath(allPoints);
