+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/toolkit/opencvtoolkit/GoalDetector.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.toolkit;
+package org.firstinspires.ftc.teamcode.toolkit.opencvtoolkit;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.opencv.core.Core;
