+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/toolkit/opencvtoolkit/RingDetectorTeleop.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.toolkit;
+package org.firstinspires.ftc.teamcode.toolkit.opencvtoolkit;
 
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
